fix config error bugs
