+++ gui/src/util/navigation.ts
@@ -1,7 +1,6 @@
 // TODO: Move all the routes here
 export const ROUTES = {
   HOME: "/",
-  CONFIG_ERROR: "/config-error",
   CONFIG: "/config",
   // EXAMPLE_ROUTE_WITH_PARAMS: (params: ParamsType) => `/route/${params}`,
 };
